macos: [perf] less IO r/w

+++ macOS/system/Dock/_dock_persistent_others.sh
@@ -1,32 +1,33 @@
 #!/bin/bash
 #
 # Dock - persistent others
+# Custom Dock's persistent `Downloads' style
 
-domain=com.apple.dock
 prefDIR=~/Library/Preferences
+domain=com.apple.dock
 plist=$prefDIR/$domain.plist
 
-# Custom Dock's persistent `Downloads' style
-persistent_others=$(plutil -extract "persistent-others" xml1 -o - "$plist")
-persistent_others=$( \
+org_persistent_others=$(plutil -extract "persistent-others" xml1 -o - "$plist")
+org_persistent_others=$( \
   echo '<dict><key>persistent-others</key><array/></dict>' | \
-  plutil -replace "persistent-others" -xml "$persistent_others" -o - - \
+  plutil -replace "persistent-others" -xml "$org_persistent_others" -o - - \
 )
+new_persistent_others="$org_persistent_others"
 
-count=$(echo "$persistent_others" | xpath "count(//array/dict)" 2>/dev/null)
+count=$(echo "$new_persistent_others" | xpath "count(//array/dict)" 2>/dev/null)
 [[ ! "$count" =~ ^[0-9]+$ ]] && count=0
 
 for (( i = 0; i < $count; i++ )); do
   entry=persistent-others.$i
   label=$( \
-    echo "$persistent_others" | \
+    echo "$new_persistent_others" | \
     plutil -extract "$entry.tile-data.file-label" xml1 -o - - | \
     plutil -p - | \
     sed -e 's/"//g' \
   )
 
   if [[ "$label" == "Downloads" ]]; then
-    setting=$(echo "$persistent_others" | plutil -extract "$entry" xml1 -o - -)
+    setting=$(echo "$new_persistent_others" | plutil -extract "$entry" xml1 -o - -)
 
     # NOTICE
     # this below command didn't execute correctly and return the error:
@@ -62,17 +63,19 @@ for (( i = 0; i < $count; i++ )); do
     # 0 => Automatic
     setting=$(echo "$setting" | plutil -replace "tile-data.showas" -integer 0 -o - -)
 
-    persistent_others=$(echo "$persistent_others" | plutil -remove "$entry" -o - -)
-    persistent_others=$(echo "$persistent_others" | plutil -replace "$entry" -xml "$setting" -o - -)
-    persistent_others=$(echo "$persistent_others" | plutil -extract "persistent-others" xml1 -o - -)
-
-    # save to Dock Preferences
-    plutil -replace "persistent-others" -xml "$persistent_others" "$plist"
+    new_persistent_others=$(echo "$new_persistent_others" | plutil -remove "$entry" -o - -)
+    new_persistent_others=$(echo "$new_persistent_others" | plutil -replace "$entry" -xml "$setting" -o - -)
+    new_persistent_others=$(echo "$new_persistent_others" | plutil -extract "persistent-others" xml1 -o - -)
 
     break
   fi
 done
 
-if [[ "$1" != "--norestart" ]]; then
-  killall Dock
+if [[ "$new_persistent_others" != "$org_persistent_others" ]]; then
+  # save to Dock Preferences
+  plutil -replace "persistent-others" -xml "$new_persistent_others" "$plist"
+
+  [[ "$1" != "--norestart" ]] && killall Dock
 fi
+
+unset new_persistent_others org_persistent_others
